+++ internal/session/session_tty_darwin.go
@@ -1,6 +1,6 @@
 //go:build darwin
 
-package main
+package session
 
 import (
 	"os"